LayoutTests/html5lib/resources/webkit02.dat

#data
<foo bar=qux/>
#errors
#document
| <html>
|   <head>
|   <body>
|     <foo>
|       bar="qux/"
